requirements.txt
gradio
numpy
Pillow
opencv-python>=4.11.0.86
setuptools==80.9.0
torch==2.9.1
torchvision==0.24.1
mani-skill @ git+https://github.com/YinpeiDai/ManiSkill.git@07be6fbc66350ddca200abfb0a11b692f078f7fd
